// gameLogd/src/app/models/documentary.model.ts
export interface Documentary {
    id: string;
    title: string;
    description: string;
    platform: string;
    releaseDate: string;
    director: string;
    productionHouse: string;
    views: number;
    rating: number;
    imageUrl: string;
    totalRatingScore: number;
    numRatings: number;
    genre: string[];
    duration: string;
    language: string[];
    subject: string;
    awards: string[];
} 